moved FollowModals outside of ProfileCard
